+++ BurningKnight/Tags.cs
@@ -30,10 +30,10 @@ namespace BurningKnight {
 		public static int ShopKeeper = new BitTag("shop_keeper");
 		public static int Npc = new BitTag("npc");
 
-		public static int Torch = new BitTag("torch");
 		public static int Button = new BitTag("button");
 		public static int Chest = new BitTag("chest");
 		public static int Projectile = new BitTag("projectile");
+		public static int MobProjectile = new BitTag("mob_projectile");
 		public static int Laser = new BitTag("laser");
 		public static int FireParticle = new BitTag("fire_particle");
 		public static int TextParticle = new BitTag("text_particle");