+++ changes.md
@@ -2,3 +2,8 @@
 * Blue bullet slime doesn't shoot after death anymore
 * Dino is a bit more angry now
 * Halo buffed
+
+* Projectile light/poof effects wont work after there are 99 projectiles in game
+* Mob projectiles are capped at 199
+* Removed Tags.Torch => Tags.MobProjectile
+* Projectiles wont be able to hit mobs/paintings/tnt in rooms that there are no players in